Everything is kinda broken.
I need to redo a few things for everything to work. Having links on
Windows is proving to be kinda difficult and I will be looking in to
having a general configuration file instead or trying to automate
everything.
You should have to set this up one time and after that "just work".